tidy up enemy-related folders
