+++ Assets/Scripts/Enemy/EnemyComponent.cs.meta
