+++ Assets/Scripts/Enemy/EnemyInterpreter.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6016d8a0c94cc7b4b9f42eddb2a9c06b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
