+++ Assets/Scripts/Enemy/EnemyInterpreter/EnemyLexer.cs
@@ -1,0 +1,7 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class EnemyLexer {
+    
+}
