+++ Assets/Scripts/Enemy/EnemyInterpreter/EnemyVM/InstructionSet/Comparison.cs
