+++ Assets/Scripts/Enemy/EnemyInterpreter/EnemyVM/Mnemonic.cs
