+++ Assets/Scripts/Enemy/EnemyInterpreter/EnemyVM/Tester/Arithmetic.cs
