+++ Assets/Scripts/Enemy/EnemyInterpreter/EnemyVM/Tester/Comparison.cs
