+++ Assets/Scripts/Enemy/EnemyInterpreter/EnemyVM/Tester/EnemyFunctionsFatoryTester.cs
