+++ Assets/Scripts/Enemy/EnemyInterpreter/EnemyVM/Tester/LogicTest.cs
