--- Assets/Scripts/EnemyInterpreterComponents/EnemyLexer.cs
@@ -1,7 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class EnemyLexer {
-    
-}
