+++ Assets/Scripts/Utils/WindowInformation.cs
