+++ Assets/Scripts/Utils/WindowInformation.cs.meta
